Use consistent failure conditions for build stages
The idea is that when a build in a stage fails, then the stage fails as well
(add a problem). And a stage fails which is a dependency of the current stage,
then the stage failed to start. For composited builds which group a test
coverage, those should also fail (add a problem), when one of their build fails.
If a dependency of an actual build (not-composite) fails, then the build also
failed to start.

Finally, when a dependency of any build failed to start, the current build
failed to start.